virerror: Introduce VIR_ERR_OPERATION_UNSAFE
Similarly to VIR_ERR_MIGRATION_UNSAFE add a error code which will be
reported when an operation is possible with the hypervisor but may lead
to data loss or other problems in certain cases. This error code
notifies the user that the operation can be forced using a specific
flag.
Message-Id: <00e979a7ae91f0180c5ab017bebf4f02cbc9d4c0.1490360622.git.pkrempa@redhat.com>