Remove usage of __attribute__((nonnull))
The attribute (defined as ATTRIBUTE_NONNULL) was added long time
ago (2009), but in 2012 (commit eefb881d4683) it was disabled for
normal build, making it used only when coverity was building libvirt
or on special request.  It was disabled because it was being misused
and misunderstood -- the attribute is there as an optimization hint
for the compiler, not to enhance static analysis.

However, it is used until today in many places and since it is
disabled by default, it just screws up builds with STATIC_ANALYSIS
enabled.

Remove that attribute and all its usages.

Moreover lv_cv_static_analysis should not be used after this patch as
it does not have any other effect than just disabling one check in
domain_conf.c.  That check is there for developers to make sure no
device is forgotten, which for coverity is just a dead code.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <4c92353ab0fe39ddac2ffc5375ec4e07ec72b137.1490701502.git.mkletzan@redhat.com>