nodedev: Introduce the mdev capability to the nodedev driver structure
Besides updating the capability enum, this patch introduces
'virNodeDevCapMdev' structure which will store everything libvirt can
gather from sysfs about a mediated device. Since we need to report the
info for both the mediated child device it's parent mdev capabilities
(merely the mdev types' attributes), this structure serves in both of
these cases with the difference being that the amount of data it holds
depends on the specific scenario (child vs parent).

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <46d6174c53c51803fe6474602555ef2d6b1ab577.1490791809.git.eskultet@redhat.com>