docs: don't install news.rng file
The news.rng file gets picked up by the wildcard match on docs/schemas,
and so gets installed into /usr/share/libvirt, along with the schemas
for our API objects. This then generated an RPM build failure, since
we don't have news.rng in the RPM file list

The news.rng schema is not something that users of libvirt ever care
about, since it is purely an internal thing to validate our website
content.

So rather than add news.rng to the RPM file list, this changes the
make rules so that we don't install news.rng at all.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
Message-Id: <20170404172407.10529-1-berrange@redhat.com>