util: Remove virsysfs and instead enhance virFileReadValue* functions
It is no longer needed thanks to the great virfilemock.c.  And this
way we don't have to add a new set of functions for each prefixed
path.

While on that, add two functions that weren't there before, string and
scaled integer reading ones.  Also increase the length of the string
being read by one to accompany for the optional newline at the
end (i.e. change INT_STRLEN_BOUND to INT_BUFSIZE_BOUND).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <ec697ff499e8604a91996f04a665e1c261aa3d7a.1491402571.git.mkletzan@redhat.com>