util: Introduce virObjectPoolableHashElement
Add a new virObjectLockable child which will be used to describe
elements that would be inserted into the virObjectPoolableHashTable.

Each virObjectPoolableHashElement will have a primaryKey and a secondaryKey
which can be used to insert the same object into both of the hash tables
that are part of the virObjectPoolableHashTable.  This allows for lookup
of the object by more than one means.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170406121052.5815-8-jferlan@redhat.com>