conf, docs: Add support for coalesce setting(s)
We are currently parsing only rx_max_coalesced_frames because that's
the only value that makes sense for us.  The tun device just added
support for this one and the others are only supported by hardware
devices which we don't need to worry about as the only way we'd pass
those to the domain is using <hostdev/> or <interface type='hostdev'/>.
And in those cases the guest can modify the settings itself.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <5d6559beb7d338ec755d4282afaf47675de4838e.1491580983.git.mkletzan@redhat.com>