qemu: Report shutdown event details
QEMU will likely report the details of it shutting down, particularly
the system it received in case it was killed.  We should forward that
information along.  Since the only way we can extend information
provided to the user is adding event details, we might as well emit
multiple events (one with the reason for the shutdown and keep the one
for the shutdown being finished for clarity and compatibility).

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1384007

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <57d4f890881275431f664fe0688ecfd96e2d8210.1492008339.git.mkletzan@redhat.com>