virfdstream: Use messages instead of pipe
One big downside of using the pipe to transfer the data is that
we can really transfer just bare data. No metadata can be carried
through unless some formatted messages are introduced. That would
be quite painful to achieve so let's use a message queue. It's
fairly easy to exchange info between threads now that iohelper is
no longer used.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <212752bbb3d49a8f3f4f1a43a1165218446a01cc.1492089792.git.mprivozn@redhat.com>