Introduce virStreamSkip
This API can be used to tell the other side of the stream to skip
some bytes in the stream. This can be used to create a sparse
file on the receiving side of a stream.

It takes just one argument @length, which says how big the hole
is. Since our streams are not rewindable like regular files, we
don't need @whence argument like seek(2) has.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <758b540a5100098aa524e8d46f0435afe4bc9ba7.1492089792.git.mprivozn@redhat.com>