Introduce virStreamInData
This is just an internal API, that calls corresponding function
in stream driver. This function will set @data=1 if the
underlying file is in data section, or @data=0 if it is in a
hole. At any rate, @length is set to number of bytes remaining in
the section the file currently is.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <93bba410ab2fe8bf9a063d242be95f2ba4bc6c71.1492089792.git.mprivozn@redhat.com>