virnetclientstream: Introduce virNetClientStreamHandleSkip
This is a function that handles an incoming STREAM_SKIP packet.
Even though it is not wired up yet, it will be soon. At the
beginning do couple of checks whether server plays nicely and
sent us a STREAM_SKIP packed only after we've enabled sparse
streams. Then decodes the message payload to see how big the hole
is and stores it in passed @length argument.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <ae970c8cf669bb309f998760c8500169eda9b80b.1492089792.git.mprivozn@redhat.com>