virsh: Implement sparse stream to vol-download
The command grew new --sparse switch that does nothing more than
enables the sparse streams feature for this command. Among with
the switch new helper function is introduced: virshStreamSkip().
This is the callback that is called whenever daemon sends us a
hole. In the callback we reflect the hole in underlying file by
seeking as many bytes as told.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <838ab73de2c8b08dd9cef4a0d7c62639e44ff29f.1492089792.git.mprivozn@redhat.com>